Continue work on our new ideas
