+++ oc/compiler/build_system/build_system.c
@@ -11,23 +11,20 @@
 static u_int32_t current_node_id = 0;
 
 /**
- * Define a node for our build graph. Every file will get
- * a build graph node that will contain it token stream
- * and record its dependencies
+ * Basic enum for the build system status
+ * That we'll pass around
  */
-typedef struct build_graph_node_t build_graph_node_t;
-struct build_graph_node_t {
-	//The token stream for the file in question
-	ollie_token_stream_t token_stream;
-	//TODO MAY UPDATE AS NEEDS ARISE
-	dynamic_array_t depends_on;
-	dynamic_array_t depended_on_by;
-	//Unique node ID
-	u_int32_t current_node_id;
-	//Name of the file that this node came from
-	char* file_name;
-};
+typedef enum {
+	BUILD_SYSTEM_STATUS_FAILURE,
+	BUILD_SYSTEM_STATUS_SUCCESS
+} build_system_status_t;
 
+/**
+ * Similar to how CFG results work, we will pass around a struct
+ * that contains our status and our created build graph nodes
+ * as a return value
+ */
+typedef struct build_system_results_t build_system_results_t;
 
 /**
  * For each file, we must:
@@ -45,46 +42,34 @@ static void handle_file_dependencies_and_tokenization(char* file_name, u_int8_t 
 
 
 /**
+ * The main file in ollie is the file that the user has passed in via the -f option
+ * to the ollie compiler
+ *
+ * The main file is a special case because it may *not* be a module. We will need to
+ * validate that the user is not attempting to make this file into a module
+ */
+static void handle_main_file_tokenization(char* main_file_name, u_int8_t silent_mode){
+	//Let's first tokenize the main file
+	ollie_token_stream_t stream = tokenize(main_file_name, silent_mode);
+
+	/**
+	 * If tokenizing failed there's no point in going further.
+	 * We fail out here and don't even bother returning anything
+	 */
+	if(stream.status == STREAM_STATUS_FAILURE){
+
+	}
+
+
+
+}
+
+
+/**
  * The main and only entry point to the build system revolves around
  * us parsing dependencies and constructing them into one gigantic, unified token
  * stream. This token stream is what we will use to actually parse and construct
  * the overall CFG
- *
- *
- * TODO $module $endmodule IDEA like we've done before with macros
- *
- * Example:
- *
- * $module "system.printer"
- *
- * //dependencies as need be
- *
- * //Actual code
- *
- * $endmodule
- *
- * Can be as big or as small as you want
- *
- * Each module should generate its own unique token stream - each module must be independently operable
- *
- * Many things to check with global vars, functions etc that need testing
- *
- * LATER IDEA NOT RIGHT NOW - submodules maybe
- *
- * BUT THEN HOW DO WE HANDLE FILES? How are we going to know which files to pull in and
- * compile? If one file can have multiple modules then how do we find the modules on the first
- * go around?
- *
- *
- * Common pattern:
- *
- * search in two areas:
- * 	The directory that we are being run in and all subdirectories under it
- * 	The common library (/usr/lib/ollie/)
- *
- * We will search through here in an attempt to find our module
- *
- * NOTE: OLLIE WILL NEVER SUPPORT INCREMENTAL BUILDS OR CACHING PRECOMP'D FILES
  */
 ollie_token_stream_t parse_dependencies_and_construct_token_stream(compiler_options_t* options, u_int8_t silent_mode){
 	/**
